+++ src/features/cart/model/useCartStore.ts
@@ -3,7 +3,7 @@ import { persist } from "zustand/middleware";
 import { SelectedOption } from "@/entities/goods/@x/cart";
 import { CartItem } from "@/entities/cart/cart.types";
 
-export type CartState = {
+type CartState = {
   items: CartItem[];
   addItem: (newItem: CartItem) => void;
   removeItem: (id: string) => void;
